refactor: auth 폴더를 apis/user 폴더로 변경
