+++ src/app/(without-side)/profile-setting/page.tsx
@@ -5,8 +5,8 @@ import Image from 'next/image';
 import Nickname from '@/components/features/signup/nickname';
 import { useEffect, useRef, useState } from 'react';
 import { TeamDto } from '@/services/apis/team/dto';
-import { UpdateUserInfoRequest } from '@/services/auth/dto';
-import { getUserInfo, updateUserInfo } from '@/services/auth';
+import { UpdateUserInfoRequest } from '@/services/apis/user/dto';
+import { getUserInfo, updateUserInfo } from '@/services/apis/user';
 import { useCurrentUserInfoStore } from '@/lib/store/useCurrentUserInfoStore';
 import { setCookie } from '@/lib/utils/cookie';
 import { getPresignedUrl, uploadToS3 } from '@/services/apis/image-upload';
