+++ src/app/(without-side)/signup/page.tsx
@@ -4,8 +4,8 @@ import Checkbox from '@/components/features/signup/checkbox';
 import Image from 'next/image';
 import { useEffect, useState } from 'react';
 import Nickname from '@/components/features/signup/nickname';
-import { UpdatePrivacyRequest, UpdateUserInfoRequest } from '@/services/auth/dto';
-import { updatePrivacy, updateUserInfo } from '@/services/auth';
+import { UpdatePrivacyRequest, UpdateUserInfoRequest } from '@/services/apis/user/dto';
+import { updatePrivacy, updateUserInfo } from '@/services/apis/user';
 import { agreementDatas } from '@/lib/constants/agreementDatas';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { getCookie, setCookie } from '@/lib/utils/cookie';
